bug fixed -
- in Object_variable_info element of stack_map_table
- in Runtime*visibleParameterAnnotations

+++ src/main/java/com/mstruzek/ccbyte/attribute/RuntimeInvisibleParameterAnnotations.java
@@ -7,7 +7,7 @@ import com.mstruzek.ccbyte.base.attribute_info;
 
 @Discriminator(attributeName = "RuntimeInvisibleParameterAnnotations")
 public class RuntimeInvisibleParameterAnnotations extends attribute_info {
-  public short num_parameters;
+  public byte num_parameters;
   @ArrayLength(fieldName = "num_parameters")
   public parameter_annotation[] parameter_annotations;
 }
